requirements.txt
discord.py>=2.5.0
openai>=1.75.0
python-dotenv>=1.0.1
